# openapi_apidog.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bimbi AI API",
    "description": "API collection documentation for Bimbi AI Backend. Built with Clean Architecture using Go and Gin. Includes authentication, RAG-based insights generation, longitudinal child tracking, and home activity recommendations with done-status tracking.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    },
    {
      "url": "http://bimbi-be-dev.btdev.my.id",
      "description": "Development Environment"
    }
  ],
  "paths": {
...
    "/api/children": {
      "get": {
        "summary": "Get All Children",
        "description": "Get a list of all child profiles belonging to the authenticated user.",
        "operationId": "getChildren",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Child" }
                },
                "example": [
                  {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "parent_id": "orang.tua@example.com",
                    "name": "Ara",
                    "birth_date": "2021-05-20T00:00:00Z",
                    "created_at": "2026-06-01T10:00:00Z",
                    "updated_at": "2026-06-01T10:00:00Z"
                  },
                  {
                    "id": "456e7890-e89b-12d3-a456-426614174001",
                    "parent_id": "orang.tua@example.com",
                    "name": "Bima",
                    "birth_date": "2019-03-15T00:00:00Z",
                    "created_at": "2026-06-02T08:00:00Z",
                    "updated_at": "2026-06-02T08:00:00Z"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "unauthorized", "message": "Missing or invalid token" }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "db_error", "message": "Failed to fetch children" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Child Profile",
        "description": "Create a new child profile for longitudinal tracking. Requires JWT authentication.",
        "operationId": "createChild",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateChildRequest" },
              "example": {
                "name": "Ara",
                "birth_date": "2021-05-20"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Child profile created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Child" },
                "example": {
                  "id": "123e4567-e89b-12d3-a456-426614174000",
                  "parent_id": "orang.tua@example.com",
                  "name": "Ara",
                  "birth_date": "2021-05-20T00:00:00Z",
                  "created_at": "2026-06-11T03:00:00Z",
                  "updated_at": "2026-06-11T03:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "invalid_payload", "message": "birth_date is required" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "unauthorized", "message": "Missing or invalid token" }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "db_error", "message": "Failed to create child" }
              }
            }
          }
        }
      }
    },
    "/api/children/{id}": {
      "get": {
        "summary": "Get Child by ID",
        "operationId": "getChild",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Child UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Child" },
                "example": {
                  "id": "123e4567-e89b-12d3-a456-426614174000",
                  "parent_id": "orang.tua@example.com",
                  "name": "Ara",
                  "birth_date": "2021-05-20T00:00:00Z",
                  "created_at": "2026-06-01T10:00:00Z",
                  "updated_at": "2026-06-01T10:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "not_found", "message": "Child not found" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "unauthorized", "message": "Missing or invalid token" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Child Profile",
        "operationId": "updateChild",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Child UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateChildRequest" },
              "example": {
                "name": "Ara Putri",
                "birth_date": "2021-05-20"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Child" },
                "example": {
                  "id": "123e4567-e89b-12d3-a456-426614174000",
                  "parent_id": "orang.tua@example.com",
                  "name": "Ara Putri",
                  "birth_date": "2021-05-20T00:00:00Z",
                  "created_at": "2026-06-01T10:00:00Z",
                  "updated_at": "2026-06-11T10:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "not_found", "message": "Child not found" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Child Profile",
        "operationId": "deleteChild",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Child UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                },
                "example": { "message": "Child deleted successfully" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "not_found", "message": "Child not found" }
              }
            }
          }
        }
      }
    },
...
    "/api/children/{id}/feedback": {
      "post": {
        "summary": "Submit Activity Feedback",
        "description": "Submit parent feedback for an activity under the child's LATEST assessment. Backend resolves the assessment automatically — no need to pass or know `assessment_id`.\n\nActivity can be:\n- A **recommended activity** from `GET /api/children/{id}/home-activities` (copy `activity_name` as-is)\n- Any **free-text activity** the parent did independently\n\nSubmitting feedback marks the matching activity as `done: true` in the home-activities list (case-insensitive match). This data is also injected into the next RAG assessment prompt to improve future recommendations.",
        "operationId": "submitFeedback",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Child UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubmitFeedbackRequest" },
              "examples": {
                "recommended_activity_struggled": {
                  "summary": "Recommended activity — struggled",
                  "value": {
                    "activity_name": "Bangun 'kota mini' dari kardus bekas sabun dan botol shampoo bersama Ara selama 15 menit sebelum tidur.",
                    "parent_experience": "Anak sangat antusias tapi kesulitan menempelkan kardusnya karena lem yang kurang kuat. Dia jadi sedikit frustrasi di akhir.",
                    "status": "struggled"
                  }
                },
                "recommended_activity_completed": {
                  "summary": "Recommended activity — completed",
                  "value": {
                    "activity_name": "Ajak Ara mendekorasi tempat belajarnya sendiri dengan gambar-gambar yang dia buat.",
                    "parent_experience": "Ara sangat senang dan fokus selama 30 menit penuh. Hasilnya bagus sekali dan dia bangga memperlihatkannya.",
                    "status": "completed"
                  }
                },
                "free_text_activity": {
                  "summary": "Free-text activity (not from AI recommendation)",
                  "value": {
                    "activity_name": "Membuat origami burung bersama kakek",
                    "parent_experience": "Ara sangat sabar mengikuti langkah-langkah lipatan kertas. Butuh 2 kali percobaan tapi akhirnya berhasil.",
                    "status": "completed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feedback saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                },
                "example": { "message": "Feedback saved successfully" }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "error": "invalid_payload",
                  "message": "Key: 'SubmitFeedbackRequest.Status' Error:Field validation for 'Status' failed on the 'oneof' tag"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "unauthorized", "message": "Missing or invalid token" }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity — no assessment found for this child yet. Run POST /api/assessments first.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "error": "service_error",
                  "message": "no assessment found for child — run an assessment first before submitting feedback"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "service_error", "message": "failed to save feedback" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token obtained from the `/api/auth/login` endpoint."
      }
    },
    "schemas": {
...
      "CreateChildRequest": {
        "type": "object",
        "required": ["name", "birth_date"],
        "properties": {
          "name": { "type": "string", "example": "Ara" },
          "birth_date": {
            "type": "string",
            "format": "date",
            "description": "Format: YYYY-MM-DD",
            "example": "2021-05-20"
          }
        }
      },
      "Child": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "68a212bf-289c-4447-aa52-1f13eeefceed" },
          "parent_id": { "type": "string", "format": "uuid", "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479" },
          "name": { "type": "string", "example": "Ara" },
          "birth_date": { "type": "string", "format": "date-time", "example": "2021-05-20T00:00:00Z" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-06-01T10:00:00Z" },
          "updated_at": { "type": "string", "format": "date-time", "example": "2026-06-01T10:00:00Z" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "invalid_payload" },
          "message": { "type": "string", "example": "Key: 'Field' Error:Field validation failed" }
        }
      },
      "SubmitFeedbackRequest": {
        "type": "object",
        "required": ["activity_name", "parent_experience", "status"],
        "properties": {
          "activity_name": {
            "type": "string",
            "description": "Name of the activity. Copy as-is from home-activities list for recommended ones, or enter free-text for custom activities.",
            "example": "Bangun 'kota mini' dari kardus bekas sabun dan botol shampoo bersama Ara selama 15 menit sebelum tidur."
          },
          "parent_experience": {
            "type": "string",
            "description": "Parent's notes about how the activity went",
            "example": "Anak sangat antusias tapi kesulitan menempelkan kardusnya karena lem yang kurang kuat. Dia jadi sedikit frustrasi di akhir."
          },
          "status": {
            "type": "string",
            "enum": ["completed", "struggled"],
            "description": "completed = activity went well; struggled = child or parent had difficulty",
            "example": "struggled"
          }
        }
      },
...
    }
  }
}